Revert "Templates"
