--- .gitignore
@@ -1,3 +1,0 @@
-.vs/
-Debug/
-x64
