--- overload.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30907.101
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "overload", "overload.vcxproj", "{E153463C-E00B-42AA-94EB-AC9EE33BA4CE}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E153463C-E00B-42AA-94EB-AC9EE33BA4CE}.Debug|x64.ActiveCfg = Debug|x64
-		{E153463C-E00B-42AA-94EB-AC9EE33BA4CE}.Debug|x64.Build.0 = Debug|x64
-		{E153463C-E00B-42AA-94EB-AC9EE33BA4CE}.Debug|x86.ActiveCfg = Debug|Win32
-		{E153463C-E00B-42AA-94EB-AC9EE33BA4CE}.Debug|x86.Build.0 = Debug|Win32
-		{E153463C-E00B-42AA-94EB-AC9EE33BA4CE}.Release|x64.ActiveCfg = Release|x64
-		{E153463C-E00B-42AA-94EB-AC9EE33BA4CE}.Release|x64.Build.0 = Release|x64
-		{E153463C-E00B-42AA-94EB-AC9EE33BA4CE}.Release|x86.ActiveCfg = Release|Win32
-		{E153463C-E00B-42AA-94EB-AC9EE33BA4CE}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E9338C96-F432-4974-8E65-E0729CD54D06}
-	EndGlobalSection
-EndGlobal
