--- overload.txt
@@ -1,8 +1,0 @@
-https://github.com/yevgeniishvedenko/overload
-
-сделал сортировку и сумму. на остальное просто времени не хватает из-за работы. но суть то понятна мне.
-кода много, можно было чуть уменьшить, но совсем чуть-чуть, добавив функции принт и филл, поэтому я это не делал.
-
-на сортировке символьного двумерного массива выдавало ошибку что ROWS и COLS не являются константами, а потом просто
-при запуске виндовс выдает звук ошибки, сама ошибка не появляется, и все работает. видимо очередной глюк студии.
-вдруг у Вас не запустится, то нужно закоментить выполнение сортировки символьного массива двумерного.
